contract_analysis/cms-ai/pyproject.toml
[project]
name = "cms-ai-claude"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "docx2txt>=0.9",
    "fastapi==0.104.1",
    "llama-index==0.10.68",
    "motor>=3.5.3",
    "notebook>=7.4.5",
    "openai==1.97.1",
    "pydantic==2.5.0",
    "pydantic-settings>=2.2.1",
    "pymongo==4.6.0",
    "pypdf2==3.0.1",
    "python-docx==1.1.0",
    "python-dotenv>=1.1.1",
    "python-multipart==0.0.6",
    "setuptools>=80.9.0",
    "uvicorn==0.24.0",
]
